move old_timestamps into walk as legacy_timestamps
