+++ picopt/walk/skip.py
@@ -7,9 +7,9 @@ from confuse import AttrDict
 from treestamps import Grovestamps, Treestamps
 
 from picopt import PROGRAM_NAME, WORKING_SUFFIX
-from picopt.old_timestamps import OLD_TIMESTAMPS_NAME
 from picopt.path import PathInfo, is_path_ignored
 from picopt.printer import Printer
+from picopt.walk.legacy_timestamps import OLD_TIMESTAMPS_NAME
 
 
 class WalkSkipper:
